+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 All notable changes to this project will be documented in this file.
 
+[1.0.5] - 2024-06-11
+ - Fixed ExtensionsCanvasGroup.cs warnings.
+ - Added custom editor script to EditorToolCanvasGroupController.cs.
+
 [1.0.4] - 2024-06-11
  - Removed ExtensionsFloat Remap warnings.